Move query builders to a cli module
